vm._len and objsequence.rs -> sliceable.rs
