+++ vm/src/builtins.rs
@@ -28,7 +28,6 @@ mod decl {
     use crate::obj::objint::{self, PyIntRef};
     use crate::obj::objiter;
     use crate::obj::objlist::{PyList, SortOptions};
-    use crate::obj::objsequence;
     use crate::obj::objstr::{PyStr, PyStrRef};
     use crate::obj::objtype::{self, PyTypeRef};
     use crate::pyobject::{
@@ -37,6 +36,7 @@ mod decl {
     };
     use crate::readline::{Readline, ReadlineResult};
     use crate::scope::Scope;
+    use crate::sliceable;
     use crate::slots::PyComparisonOp;
     #[cfg(feature = "rustpython-parser")]
     use crate::stdlib::ast;
@@ -444,7 +444,12 @@ mod decl {
 
     #[pyfunction]
     fn len(obj: PyObjectRef, vm: &VirtualMachine) -> PyResult<usize> {
-        objsequence::len(&obj, vm)
+        vm._len(&obj).unwrap_or_else(|| {
+            Err(vm.new_type_error(format!(
+                "object of type '{}' has no len()",
+                obj.lease_class().name
+            )))
+        })
     }
 
     #[pyfunction]
